+++ app/models/cms/html_block.rb
@@ -11,7 +11,6 @@ module Cms
     # Override of search scope from searching behavior to deal with include_body
     scope :search, lambda { |search_params|
       term = search_params.is_a?(Hash) ? search_params[:term] : search_params
-      order = search_params.is_a?(Hash) && search_params[:order] ? search_params[:order] : table_name.to_s + ".name"
       include_body = search_params.is_a?(Hash) ? search_params[:include_body] : false
       conditions = []
       columns = ["name"]
@@ -29,7 +28,6 @@ module Cms
       end
       scope = {}
       scope[:conditions] = conditions if conditions
-      scope[:order] = order if order
       scope
     }